requirements.txt
pandas
numpy
requests
sqlalchemy
psycopg2-binary
apache-airflow
python-dotenv
pytest